Stats (#1143)
* Looks at stats of cycle spectra

* Updates to fix several small bugs

Problm with generating a coordinate system when two input vectors parallel
or anti-parallel #1141

The problem reported in #1127 regarding run_check.py failing
when there are multiple domains. The routine no longer fails
but this is not a complete fix, since the results that are
reported only refer to the 0 domain.  A more complete fix seems
unjustified because multiple domains are rarely used.