Updates to switch to use bands in cells
The code has now been modified so that
estimators are calculated in the cell
bands, ranther than the photon generation
bands.

At present, the limits are the same in
both places but this can now be changed
if desired.

Note that there is an open question
regarding the banding that is currently
transmitted in the rad hydro routines