Cleanup based on valgrind
A number of small issues noted
by valgrind have been fixed.

Note that some of the rdpar inputs
do not have the proper last character.
A change I made looks for this problem,
and adds the EOL character, but it's
unclear why some inputs have this problem.

Also, found in fb, that were accessing one
to many array elements in the density
arrays.