Xnew (#1159)
This adds a new ionization method that allows one to run one_shot multiple times.

It changes when photmain is a allocated and then freed, With this change, space for photmain is allocated at the beginging of photon generation and deallocated after all estimators have been accumulated, but before the actual ionization changes are changed. The goal os this change is to minimize the maximum memory required by sirocco.

In the process of making these changes a number of issues came up, having to do principly with overflows by one, of some arrays, including photmain. These were fixed by increasing the size of photmain by 1, but it is not entirely obvious that this "bandaid" does not obscure some kind of other mismatch.

Valgrind also revealed that some of the rdpar inputs are not terminated. A check has been added for this, and the problem has been corrected, but I have not investigated why some inputs have this problem and others do not. An error is generated for the problems that do have the problems.

There is now additional information produced of the memory being used by sirocco during various parts of the program. This information is more extensive on linux than mac_os.

The version number has been incremented, because there are some changes in the windsave files; which means that this version of Sirocco does not read old versions of windsave files.