Fix string ndings in string2int (#1162)
This change properly terminates strings created in string2int which is part of rdpar.c, and avoids throwing errors while fixing the problem later in the code.

Checks to see that the strings are properly terminated are still made, but should not generate errors.  

(Note that improperly terminated strings generate warnings/errors in Valgrind because they suggest that one may be accessing memory which is part of the string.)